backend/foodgram/requirements.txt
asgiref==3.5.2
backports.zoneinfo==0.2.1
certifi==2022.6.15
cffi==1.15.1
charset-normalizer==2.1.1
coreapi==2.3.3
coreschema==0.0.4
django-filter==22.1
django-more-admin-filters==1.3
django-templated-mail==1.1.1
djangorestframework==3.12.4
djangorestframework-simplejwt==4.8.0
cryptography==37.0.4
defusedxml==0.7.1
Django==3.2.15
django-admin-autocomplete-filter==0.7.1
djoser==2.1.0
drf-extra-fields==3.4.0
idna==3.3
itypes==1.2.0
Jinja2==3.1.2
MarkupSafe==2.1.1
oauthlib==3.2.0
Pillow==8.3.1
pycparser==2.21
six==1.16.0
social-auth-app-django==4.0.0
social-auth-core==4.3.0
sqlparse==0.4.2
uritemplate==4.1.1
PyJWT==2.4.0
python3-openid==3.2.0
pytz==2022.2.1
requests==2.28.1
requests-oauthlib==1.3.1
urllib3==1.26.12
gunicorn==20.0.4
psycopg2-binary==2.8.6